+++ lmcache/v1/memory_allocators/buffer_allocator.py
@@ -1,0 +1,75 @@
+# SPDX-License-Identifier: Apache-2.0
+
+# Standard
+from typing import List, Optional, Union
+
+# Third Party
+import torch
+
+# First Party
+from lmcache.utils import _lmcache_nvtx_annotate
+from lmcache.v1.memory_management import (
+    BytesBufferMemoryObj,
+    MemoryAllocatorInterface,
+    MemoryFormat,
+    MemoryObj,
+)
+
+
+class BufferAllocator(MemoryAllocatorInterface):
+    """Allocates memory in the pre-allocated pinned memory."""
+
+    def __init__(self, device="cpu"):
+        """
+        :param str device: The device of the buffer memory.
+        """
+        self.device = device
+
+    @_lmcache_nvtx_annotate
+    def allocate(
+        self,
+        shapes: Union[torch.Size, list[torch.Size]],
+        dtypes: Union[torch.dtype, list[torch.dtype]],
+        fmt: MemoryFormat = MemoryFormat.BINARY_BUFFER,
+        allocator_type: Optional[str] = None,
+    ) -> BytesBufferMemoryObj:
+        if isinstance(shapes, list):
+            n = shapes[0][0]
+        else:
+            n = shapes[0]
+        byte_array = bytearray(n)
+        return BytesBufferMemoryObj(byte_array)
+
+    @_lmcache_nvtx_annotate
+    def batched_allocate(
+        self,
+        shapes: Union[torch.Size, list[torch.Size]],
+        dtypes: Union[torch.dtype, list[torch.dtype]],
+        batch_size: int,
+        fmt: MemoryFormat = MemoryFormat.BINARY_BUFFER,
+        allocator_type: Optional[str] = None,
+    ) -> List[BytesBufferMemoryObj]:
+        if isinstance(shapes, list):
+            n = shapes[0][0]
+        else:
+            n = shapes[0]
+        # TODO(Jiayi): Optimize the following loop.
+        byte_arrays = [bytearray(n) for _ in range(batch_size)]
+        return [BytesBufferMemoryObj(byte_array) for byte_array in byte_arrays]
+
+    def free(self, memory_obj: MemoryObj, allocator_type: Optional[str] = None):
+        return
+
+    def batched_free(
+        self,
+        memory_objs: List[MemoryObj],
+        allocator_type: Optional[str] = None,
+        update_stats: bool = True,
+    ):
+        return
+
+    def __str__(self):
+        return "BufferAllocator"
+
+    def memcheck(self):
+        return True
