+++ lmcache/v1/memory_management.py
@@ -1,14 +1,11 @@
 # SPDX-License-Identifier: Apache-2.0
 # Standard
-from collections import deque
-from contextlib import nullcontext
 from dataclasses import dataclass
 from enum import Enum, auto
 from functools import cache, wraps
 from typing import Any, List, Optional, Tuple, Union
 import abc
 import ctypes
-import mmap
 import os
 import threading
 
@@ -1611,1738 +1608,43 @@ class AddressManager:
         return True
 
 
-class TensorMemoryAllocator(MemoryAllocatorInterface):
-    """
-    Implements a "explicit list" memory allocator.
-    Uses AddressManager for address space management.
-    """
-
-    def __init__(
-        self,
-        tensor: torch.Tensor,
-        align_bytes: int = AddressManager.ALIGN_BYTES,
-        init_address_space: int | None = None,
-    ):
-        """
-        Args:
-            tensor: The pre-allocated flat tensor to use as the memory pool.
-            align_bytes: The alignment requirement for allocations.
-            init_address_space: Initial size of the address space. If None,
-                use the size of the provided tensor.
-
-        Note:
-            The `init_address_space` is used for lazy memory allocation.
-            We probably want to have a better way to make sure that the
-            LazyMemoryAllocator can be decoupled from TensorMemoryAllocator.
-        """
-        self.buffer = tensor.view(torch.uint8).flatten()
-
-        # Use AddressManager for address space management
-        self.address_manager = AddressManager(
-            self.buffer.numel() if init_address_space is None else init_address_space,
-            align_bytes,
-        )
-
-        # For debugging purposes
-        self.num_active_allocations = 0
-
-        self.stats_monitor = LMCStatsMonitor.GetOrCreate()
-
-    @property
-    def total_allocated_size(self) -> int:
-        return self.address_manager.total_allocated_size
-
-    @_lmcache_nvtx_annotate
-    def allocate(
-        self,
-        shapes: Union[torch.Size, list[torch.Size]],
-        dtypes: Union[torch.dtype, list[torch.dtype]],
-        fmt: MemoryFormat = MemoryFormat.KV_2LTD,
-        allocator_type: Optional[str] = None,
-    ) -> Optional[TensorMemoryObj]:
-        shapes, dtypes = self._adapt_shapes_and_dtypes(shapes, dtypes)
-
-        # Calculate the size of the tensor
-        raw_size = get_size_bytes(shapes, dtypes)
-
-        # Allocate from address manager
-        try:
-            block_start, aligned_size = self.address_manager.allocate(raw_size)
-        except RuntimeError:
-            # No block found
-            return None
-
-        # For debug
-        self.num_active_allocations += 1
-
-        # Update stats
-        self.stats_monitor.update_local_cache_usage(
-            self.address_manager.total_allocated_size
-        )
-        self.stats_monitor.update_active_memory_objs_count(self.num_active_allocations)
-
-        # Allocate the block
-        raw_data = self._get_buffer_slice(block_start, raw_size)
-        return TensorMemoryObj(
-            raw_data=raw_data,
-            metadata=MemoryObjMetadata(
-                shapes[0],
-                dtypes[0],
-                block_start,
-                aligned_size,
-                1,
-                0,
-                fmt,
-                shapes=shapes,
-                dtypes=dtypes,
-            ),
-            parent_allocator=self,
-        )
-
-    def _get_buffer_slice(self, start: int, size: int) -> torch.Tensor:
-        """Hook: Get buffer slice. Override for custom buffer access."""
-        return self.buffer[start : start + size]
-
-    @_lmcache_nvtx_annotate
-    def batched_allocate(
-        self,
-        shapes: Union[torch.Size, list[torch.Size]],
-        dtypes: Union[torch.dtype, list[torch.dtype]],
-        batch_size: int,
-        fmt: MemoryFormat = MemoryFormat.KV_2LTD,
-        allocator_type: Optional[str] = None,
-    ) -> Optional[List[TensorMemoryObj]]:
-        """
-        Batched allocate tensor memory objs with equal sizes.
-        """
-        shapes, dtypes = self._adapt_shapes_and_dtypes(shapes, dtypes)
-
-        # Calculate the size of the tensor
-        unit_raw_size = get_size_bytes(shapes, dtypes)
-        unit_aligned_size = self.address_manager.compute_aligned_size(unit_raw_size)
-
-        try:
-            alloc_results = self.address_manager.batched_allocate(
-                unit_aligned_size, batch_size
-            )
-        except RuntimeError:
-            return None
-        addresses = [addr for addr, _ in alloc_results]
-        raw_datas = [
-            self._get_buffer_slice(addr, unit_aligned_size) for addr in addresses
-        ]
-
-        # For debug
-        self.num_active_allocations += batch_size
-
-        # Update stats
-        self.stats_monitor.update_local_cache_usage(
-            self.address_manager.total_allocated_size
-        )
-        self.stats_monitor.update_active_memory_objs_count(self.num_active_allocations)
-
-        tensor_mem_objs = []
-        for raw_data, address in zip(raw_datas, addresses, strict=True):
-            tensor_mem_objs.append(
-                TensorMemoryObj(
-                    raw_data=raw_data,
-                    metadata=MemoryObjMetadata(
-                        shapes[0],
-                        dtypes[0],
-                        address,
-                        unit_aligned_size,
-                        1,
-                        0,
-                        fmt,
-                        shapes=shapes,
-                        dtypes=dtypes,
-                    ),
-                    parent_allocator=self,
-                )
-            )
-
-        return tensor_mem_objs
-
-    @_lmcache_nvtx_annotate
-    def free(self, memory_obj: MemoryObj, allocator_type: Optional[str] = None):
-        if not memory_obj.is_valid():
-            return
-
-        self.address_manager.free(memory_obj.meta.address, memory_obj.meta.phy_size)
-        memory_obj.invalidate()
-
-        # For debug
-        self.num_active_allocations -= 1
-
-        # Update stats
-        self.stats_monitor.update_local_cache_usage(
-            self.address_manager.total_allocated_size
-        )
-        self.stats_monitor.update_active_memory_objs_count(self.num_active_allocations)
-
-    @_lmcache_nvtx_annotate
-    def batched_free(
-        self,
-        memory_objs: List[MemoryObj],
-        allocator_type: Optional[str] = None,
-        update_stats: bool = True,
-    ):
-        """
-        Batched free memory objs.
-        Unlike `batched_allocate`, this function does not
-        assume that the memory objs are equal-sized.
-        """
-        if not memory_objs:
-            return
-
-        # Coalesce adjacent memory objects before freeing to reduce
-        # the number of free operations
-        coalesced_blocks: list[tuple[int, int, int]] = []  # (address, size, count)
-        curr_start = None
-        curr_size = 0
-        curr_count = 0
-
-        memory_objs.sort(key=lambda x: x.meta.address)
-        for memory_obj in memory_objs:
-            if not memory_obj.is_valid():
-                logger.warning("Trying to free an invalidated MemoryObj")
-                continue
-            memory_obj.invalidate()
-
-            if curr_start is None:
-                curr_start = memory_obj.meta.address
-                curr_size = memory_obj.meta.phy_size
-                curr_count = 1
-            elif curr_start + curr_size == memory_obj.meta.address:
-                # Adjacent block, extend current
-                curr_size += memory_obj.meta.phy_size
-                curr_count += 1
-            else:
-                # Non-adjacent, save current and start new
-                coalesced_blocks.append((curr_start, curr_size, curr_count))
-                curr_start = memory_obj.meta.address
-                curr_size = memory_obj.meta.phy_size
-                curr_count = 1
-
-        if curr_start is not None:
-            coalesced_blocks.append((curr_start, curr_size, curr_count))
-
-        # Free all coalesced blocks
-        total_count = 0
-        for address, size, count in coalesced_blocks:
-            self.address_manager.free(address, size)
-            total_count += count
-
-        # For debug
-        self.num_active_allocations -= total_count
-
-        if update_stats:
-            self.stats_monitor.update_local_cache_usage(
-                self.address_manager.total_allocated_size
-            )
-            self.stats_monitor.update_active_memory_objs_count(
-                self.num_active_allocations
-            )
-
-    def memcheck(self):
-        """For debug purposes.
-        Returns True is everything is fine, otherwise False.
-        """
-        clear = True
-        logger.info("Checking memory allocator consistency")
-        logger.info(" - Total active allocations: %d", self.num_active_allocations)
-        logger.info(
-            " - Total allocated size: %f MB",
-            self.address_manager.total_allocated_size / 1048576,
-        )
-
-        # Check the real total free size
-        total_free_size = self.address_manager.get_free_size()
-        logger.info(" - Total free size: %f MB", total_free_size / 1048576)
-
-        # Check if the numbers are consistent
-        if (
-            total_free_size + self.address_manager.total_allocated_size
-            != self.address_manager.get_heap_size()
-        ):
-            logger.error("Memory allocator size is inconsistent")
-            logger.error("This implies a bug in the memory allocator")
-            clear = False
-
-        # Check if the blocks are coalesced
-        if not self.address_manager.check_consistency():
-            logger.error("Memory allocator has non-coalesced blocks")
-            logger.error("This implies a bug in the memory allocator")
-            clear = False
-
-        return clear
-
-    def __str__(self):
-        return "TensorMemoryAllocator"
-
-
-class PagedAddressManager:
-    """
-    A lightweight address manager for PagedTensorMemoryAllocator.
-    Provides get_free_size() and get_heap_size() by reading the
-    paged allocator's state.
-    """
-
-    def __init__(self, paged_allocator: "PagedTensorMemoryAllocator"):
-        self._allocator = paged_allocator
-
-    def get_heap_size(self) -> int:
-        """Get the total size of the paged address space in bytes."""
-        return self._allocator.buffer_size
-
-    def get_free_size(self) -> int:
-        """Get the total free size in bytes."""
-        return len(self._allocator.free_blocks) * self._allocator.align_bytes
-
-
-class PagedTensorMemoryAllocator(MemoryAllocatorInterface):
-    """
-    Implements a paged memory allocator.
-    """
-
-    def __init__(
-        self,
-        tensor: torch.Tensor,
-        shapes: list[torch.Size],
-        dtypes: list[torch.dtype],
-        fmt: MemoryFormat = MemoryFormat.KV_2LTD,
-    ):
-        self.buffer = tensor.view(torch.uint8).flatten()
-        self.buffer_size = self.buffer.numel() * self.buffer.element_size()
-        self.buffer_ptr = self.buffer.data_ptr()
-
-        self.shapes = shapes
-        self.dtypes = dtypes
-        self.fmt = fmt
-
-        # full chunk size bytes
-        self.align_bytes = get_size_bytes(shapes, dtypes)
-
-        assert self.buffer_size % self.align_bytes == 0, (
-            f"Buffer size {self.buffer_size} must be a"
-            f" multiple of align bytes {self.align_bytes}"
-            " in paged memory allocator."
-        )
-
-        self.paged_buffers = torch.split(self.buffer, self.align_bytes, dim=0)
-
-        # NOTE: deque is used since thread-safety is not a concern here as
-        # is implemented in C under the hood (in CPython), and operations
-        # on deque are atomic.
-        self.free_blocks: deque[TensorMemoryObj] = deque()
-
-        for idx, buf in enumerate(self.paged_buffers):
-            # NOTE: idx is the paged index
-            # NOTE: the last unfull chunk's shape needs to be
-            # adjusted during allocation.
-            metadata = MemoryObjMetadata(
-                self.shapes[0],
-                self.dtypes[0],
-                idx,
-                self.align_bytes,  # 1 page
-                1,  # ref_count=1
-                0,  # pin_count=0
-                self.fmt,
-                shapes=self.shapes,
-                dtypes=self.dtypes,
-            )
-            mem_obj = TensorMemoryObj(
-                raw_data=buf,
-                metadata=metadata,
-                parent_allocator=self,
-            )
-            self.free_blocks.append(mem_obj)
-
-        # Address manager for memory usage tracking
-        self.address_manager = PagedAddressManager(self)
-
-        # For debugging purposes
-        self.num_active_allocations = 0
-        self.total_allocated_size = 0
-
-        self.stats_monitor = LMCStatsMonitor.GetOrCreate()
-        logger.info(
-            "Paged tensor memory allocator initialized, "
-            "shapes: %s, dtypes: %s, align bytes: %s",
-            self.shapes,
-            self.dtypes,
-            self.align_bytes,
-        )
-
-    @_lmcache_nvtx_annotate
-    def allocate(
-        self,
-        shapes: Union[torch.Size, list[torch.Size]],
-        dtypes: Union[torch.dtype, list[torch.dtype]],
-        fmt: MemoryFormat = MemoryFormat.KV_2LTD,
-        allocator_type: Optional[str] = None,
-    ) -> Optional[TensorMemoryObj]:
-        shapes, dtypes = self._adapt_shapes_and_dtypes(shapes, dtypes)
-
-        try:
-            free_block = self.free_blocks.popleft()
-        except IndexError:
-            logger.debug(
-                f"Failed to allocate memory for "
-                f"tensor({shapes}, {dtypes}) because "
-                "no free blocks is available"
-            )
-            return None
-
-        # TODO (Jiayi): This is a bit redundant.
-        free_block.meta.shape = shapes[0]
-        free_block.meta.dtype = dtypes[0]
-        free_block.meta.shapes = shapes
-        free_block.meta.dtypes = dtypes
-        free_block.meta.fmt = fmt
-        free_block.meta.ref_count = 1
-        # Reset any narrowed-size override left over from the previous
-        # owner of this block, so get_size() returns the layout-derived
-        # size for the fresh allocation.
-        free_block._used_size_override = None
-
-        if shapes != self.shapes:
-            size_in_bytes = get_size_bytes(shapes, dtypes)
-            free_block.raw_data = free_block.raw_data[:size_in_bytes]
-
-        # TODO (Jiayi): need a flag to drop these debug ops
-        # NOTE (Jiayi): the following code is not thread-safe but
-        # is tolerable as this is only used for debugging purposes.
-        # Update debug status
-        self.num_active_allocations += 1
-        self.total_allocated_size += self.align_bytes
-        self.stats_monitor.update_local_cache_usage(self.total_allocated_size)
-        self.stats_monitor.update_active_memory_objs_count(self.num_active_allocations)
-
-        # Allocate the block
-        return free_block
-
-    @_lmcache_nvtx_annotate
-    def batched_allocate(
-        self,
-        shapes: Union[torch.Size, list[torch.Size]],
-        dtypes: Union[torch.dtype, list[torch.dtype]],
-        batch_size: int,
-        fmt: MemoryFormat = MemoryFormat.KV_2LTD,
-        allocator_type: Optional[str] = None,
-    ) -> Optional[List[TensorMemoryObj]]:
-        """
-        Batched allocate tensor memory objs with pre-defined equal sizes.
-        """
-        shapes, dtypes = self._adapt_shapes_and_dtypes(shapes, dtypes)
-
-        allocated_blocks: list[TensorMemoryObj] = []
-        for i in range(batch_size):
-            try:
-                free_block = self.free_blocks.popleft()
-            except IndexError:
-                logger.debug(
-                    f"Failed to allocate memory for "
-                    f"tensor({shapes}, {dtypes}) because "
-                    "no free blocks is available"
-                )
-                self.batched_free(allocated_blocks, update_stats=False)
-                return None
-
-            # FIXME: think about whether parent_allocator
-            # should be updated here.
-            free_block.meta.shape = shapes[0]
-            free_block.meta.dtype = dtypes[0]
-            free_block.meta.shapes = shapes
-            free_block.meta.dtypes = dtypes
-            free_block.meta.fmt = fmt
-            free_block.meta.ref_count = 1
-            # Reset narrowed-size override (see notes in ``allocate``).
-            free_block._used_size_override = None
-
-            if shapes != self.shapes:
-                size_in_bytes = get_size_bytes(shapes, dtypes)
-                free_block.raw_data = free_block.raw_data[:size_in_bytes]
-
-            allocated_blocks.append(free_block)
-
-        # TODO (Jiayi): need a flag to drop these debug ops
-        # NOTE (Jiayi): the following code is not thread-safe but
-        # is tolerable as this is only used for debugging purposes.
-        # Update debug status
-        self.num_active_allocations += batch_size
-        self.total_allocated_size = self.num_active_allocations * self.align_bytes
-        self.stats_monitor.update_local_cache_usage(self.total_allocated_size)
-        self.stats_monitor.update_active_memory_objs_count(self.num_active_allocations)
-
-        # Allocate the block
-        return allocated_blocks
-
-    @_lmcache_nvtx_annotate
-    def free(self, memory_obj: TensorMemoryObj, allocator_type: Optional[str] = None):
-        if not memory_obj.is_valid():
-            return
-        if memory_obj.meta.shapes != self.shapes:
-            page_idx = memory_obj.meta.address
-            memory_obj.raw_data = self.paged_buffers[page_idx]
-
-        self.free_blocks.append(memory_obj)
-
-        # memory_obj.invalidate()
-
-        # TODO (Jiayi): need a flag to drop these debug ops
-        # NOTE (Jiayi): the following code is not thread-safe but
-        # is tolerable as this is only used for debugging purposes.
-        # Update debug status
-        self.total_allocated_size -= self.align_bytes
-        self.num_active_allocations -= 1
-        self.stats_monitor.update_local_cache_usage(self.total_allocated_size)
-        self.stats_monitor.update_active_memory_objs_count(self.num_active_allocations)
-
-    @_lmcache_nvtx_annotate
-    def batched_free(
-        self,
-        memory_objs: List[TensorMemoryObj],
-        allocator_type: Optional[str] = None,
-        update_stats: bool = True,
-    ):
-        """
-        Batched free memory objs.
-        Unlike `batched_allocate`, this function does not
-        assume that the memory objs are equal-sized.
-        """
-        if not memory_objs:
-            return
-
-        for memory_obj in memory_objs:
-            if not memory_obj.is_valid():
-                logger.warning("Trying to free an invalidated MemoryObj")
-                continue
-            # memory_obj.invalidate()
-            if memory_obj.meta.shapes != self.shapes:
-                page_idx = memory_obj.meta.address
-                memory_obj.raw_data = self.paged_buffers[page_idx]
-
-            self.free_blocks.append(memory_obj)
-
-        if update_stats:
-            num_freed_blocks = len(memory_objs)
-            # TODO (Jiayi): need a flag to drop these debug ops
-            # NOTE (Jiayi): the following code is not thread-safe but
-            # is tolerable as this is only used for debugging purposes.
-            # Update debug status
-            self.total_allocated_size -= self.align_bytes * num_freed_blocks
-            self.num_active_allocations -= num_freed_blocks
-            self.stats_monitor.update_local_cache_usage(self.total_allocated_size)
-            self.stats_monitor.update_active_memory_objs_count(
-                self.num_active_allocations
-            )
-
-    def memcheck(self):
-        """For debug purposes.
-        Returns True is everything is fine, otherwise False.
-        """
-
-        logger.info("Checking memory allocator consistency")
-        logger.info(" - Total active allocations: %d", self.num_active_allocations)
-        logger.info(
-            " - Total allocated size: %f MB", self.total_allocated_size / 1048576
-        )
-
-        # Check the real total free size
-        total_free_size = len(self.free_blocks) * self.align_bytes
-        logger.info(" - Total free size: %f MB", total_free_size / 1048576)
-
-        # Check if the numbers are consistent
-        if total_free_size + self.total_allocated_size != self.buffer.numel():
-            logger.error("Memory allocator size is inconsistent")
-            logger.error("This implies a bug in the memory allocator")
-            return False
-
-        return True
-
-    def __str__(self):
-        return "PagedTensorMemoryAllocator"
-
-    def get_paged_buffers(self) -> tuple[torch.Tensor, ...]:
-        """
-        Get the paged buffers for fixed buffer registration.
-
-        Returns:
-            Tuple of paged buffer tensors that can be registered with io_uring
-            for true zero copy operations.
-        """
-        return self.paged_buffers
-
-    def __del__(self):
-        # FIXME: NIXL-related memory leak should be handled somewhere (else).
-        del self.buffer
-
-
-class BufferAllocator(MemoryAllocatorInterface):
-    """Allocates memory in the pre-allocated pinned memory."""
-
-    def __init__(self, device="cpu"):
-        """
-        :param str device: The device of the buffer memory.
-        """
-        self.device = device
-
-    @_lmcache_nvtx_annotate
-    def allocate(
-        self,
-        shapes: Union[torch.Size, list[torch.Size]],
-        dtypes: Union[torch.dtype, list[torch.dtype]],
-        fmt: MemoryFormat = MemoryFormat.BINARY_BUFFER,
-        allocator_type: Optional[str] = None,
-    ) -> BytesBufferMemoryObj:
-        if isinstance(shapes, list):
-            n = shapes[0][0]
-        else:
-            n = shapes[0]
-        byte_array = bytearray(n)
-        return BytesBufferMemoryObj(byte_array)
-
-    @_lmcache_nvtx_annotate
-    def batched_allocate(
-        self,
-        shapes: Union[torch.Size, list[torch.Size]],
-        dtypes: Union[torch.dtype, list[torch.dtype]],
-        batch_size: int,
-        fmt: MemoryFormat = MemoryFormat.BINARY_BUFFER,
-        allocator_type: Optional[str] = None,
-    ) -> List[BytesBufferMemoryObj]:
-        if isinstance(shapes, list):
-            n = shapes[0][0]
-        else:
-            n = shapes[0]
-        # TODO(Jiayi): Optimize the following loop.
-        byte_arrays = [bytearray(n) for _ in range(batch_size)]
-        return [BytesBufferMemoryObj(byte_array) for byte_array in byte_arrays]
-
-    def free(self, memory_obj: MemoryObj, allocator_type: Optional[str] = None):
-        return
-
-    def batched_free(
-        self,
-        memory_objs: List[MemoryObj],
-        allocator_type: Optional[str] = None,
-        update_stats: bool = True,
-    ):
-        return
-
-    def __str__(self):
-        return "BufferAllocator"
-
-    def memcheck(self):
-        return True
-
-
-class HostMemoryAllocator(MemoryAllocatorInterface):
-    """Allocates memory in the pre-allocated Host memory."""
-
-    def __init__(self, size: int, use_paging: bool = False, **kwargs):
-        """
-        :param int size: The size of the pinned memory in bytes.
-        """
-        buffer = torch.empty(size, dtype=torch.uint8, device="cpu")
-
-        self.allocator: MemoryAllocatorInterface
-        if use_paging:
-            assert "shapes" in kwargs, (
-                "shapes must be specified for paged memory allocator"
-            )
-            assert "dtypes" in kwargs, (
-                "dtypes must be specified for paged memory allocator"
-            )
-            assert "fmt" in kwargs, "fmt must be specified for paged memory allocator"
-            self.allocator = PagedTensorMemoryAllocator(
-                tensor=buffer,
-                shapes=kwargs["shapes"],
-                dtypes=kwargs["dtypes"],
-                fmt=kwargs["fmt"],
-            )
-        else:
-            self.allocator = TensorMemoryAllocator(buffer)
-
-        self.host_mem_lock = threading.Lock() if not use_paging else nullcontext()
-
-    @_lmcache_nvtx_annotate
-    def allocate(
-        self,
-        shapes: Union[torch.Size, list[torch.Size]],
-        dtypes: Union[torch.dtype, list[torch.dtype]],
-        fmt: MemoryFormat = MemoryFormat.KV_2LTD,
-        allocator_type: Optional[str] = None,
-    ) -> Optional[MemoryObj]:
-        with self.host_mem_lock:
-            return self.allocator.allocate(shapes, dtypes, fmt, str(self))
-
-    @_lmcache_nvtx_annotate
-    def batched_allocate(
-        self,
-        shapes: Union[torch.Size, list[torch.Size]],
-        dtypes: Union[torch.dtype, list[torch.dtype]],
-        batch_size: int,
-        fmt: MemoryFormat = MemoryFormat.KV_2LTD,
-        allocator_type: Optional[str] = None,
-    ) -> Optional[List[MemoryObj]]:
-        with self.host_mem_lock:
-            return self.allocator.batched_allocate(
-                shapes, dtypes, batch_size, fmt, str(self)
-            )
-
-    @_lmcache_nvtx_annotate
-    def free(self, memory_obj: MemoryObj, allocator_type: Optional[str] = None):
-        with self.host_mem_lock:
-            self.allocator.free(memory_obj)
-
-    @_lmcache_nvtx_annotate
-    def batched_free(
-        self,
-        memory_objs: List[MemoryObj],
-        allocator_type: Optional[str] = None,
-        update_stats: bool = True,
-    ):
-        with self.host_mem_lock:
-            self.allocator.batched_free(memory_objs)
-
-    def memcheck(self):
-        with self.host_mem_lock:
-            return self.allocator.memcheck()
-
-    def __str__(self):
-        return "HostMemoryAllocator"
-
-
-class PinMemoryAllocator(MemoryAllocatorInterface):
-    """Allocates memory in the pre-allocated pinned memory."""
-
-    def __init__(self, size: int, use_paging: bool = False, **kwargs):
-        """
-        :param int size: The size of the pinned memory in bytes.
-        """
-        self.size = size
-        self.buffer = _allocate_cpu_memory(size)
-        self._unregistered = False
-
-        self.allocator: MemoryAllocatorInterface
-        if use_paging:
-            assert "shapes" in kwargs, (
-                "shapes must be specified for paged memory allocator"
-            )
-            assert "dtypes" in kwargs, (
-                "dtypes must be specified for paged memory allocator"
-            )
-            assert "fmt" in kwargs, "fmt must be specified for paged memory allocator"
-            self.allocator = PagedTensorMemoryAllocator(
-                tensor=self.buffer,
-                shapes=kwargs["shapes"],
-                dtypes=kwargs["dtypes"],
-                fmt=kwargs["fmt"],
-            )
-        else:
-            self.allocator = TensorMemoryAllocator(self.buffer)
-
-        self.host_mem_lock = threading.Lock() if not use_paging else nullcontext()
-
-    @_lmcache_nvtx_annotate
-    def allocate(
-        self,
-        shapes: Union[torch.Size, list[torch.Size]],
-        dtypes: Union[torch.dtype, list[torch.dtype]],
-        fmt: MemoryFormat = MemoryFormat.KV_2LTD,
-        allocator_type: Optional[str] = None,
-    ) -> Optional[MemoryObj]:
-        with self.host_mem_lock:
-            return self.allocator.allocate(shapes, dtypes, fmt, str(self))
-
-    @_lmcache_nvtx_annotate
-    def batched_allocate(
-        self,
-        shapes: Union[torch.Size, list[torch.Size]],
-        dtypes: Union[torch.dtype, list[torch.dtype]],
-        batch_size: int,
-        fmt: MemoryFormat = MemoryFormat.KV_2LTD,
-        allocator_type: Optional[str] = None,
-    ) -> Optional[List[MemoryObj]]:
-        with self.host_mem_lock:
-            return self.allocator.batched_allocate(
-                shapes, dtypes, batch_size, fmt, str(self)
-            )
-
-    @_lmcache_nvtx_annotate
-    def free(self, memory_obj: MemoryObj, allocator_type: Optional[str] = None):
-        with self.host_mem_lock:
-            self.allocator.free(memory_obj)
-
-    @_lmcache_nvtx_annotate
-    def batched_free(
-        self,
-        memory_objs: List[MemoryObj],
-        allocator_type: Optional[str] = None,
-        update_stats: bool = True,
-    ):
-        with self.host_mem_lock:
-            self.allocator.batched_free(memory_objs)
-
-    def memcheck(self):
-        with self.host_mem_lock:
-            return self.allocator.memcheck()
-
-    def close(self):
-        if not self._unregistered:
-            if self.buffer.numel() == 0:
-                return
-            _free_cpu_memory(self.buffer, self.size)
-            self._unregistered = True
-
-    def __str__(self):
-        return "PinMemoryAllocator"
-
-
-class MixedMemoryAllocator(MemoryAllocatorInterface):
-    """
-    Allocates (1) memory in the pre-allocated pinned memory.
-              (2) byte_array buffer memory.
-    """
-
-    def __init__(
-        self, size: int, use_paging: bool = False, use_hugepages: bool = False, **kwargs
-    ):
-        """
-        :param int size: The size of the pinned memory in bytes.
-        :param bool use_hugepages: Whether to use hugepages.
-        """
-
-        self.numa_mapping = kwargs.get("numa_mapping", None)
-        self.use_hugepages = use_hugepages
-        self.align_bytes = kwargs.get("align_bytes", AddressManager.ALIGN_BYTES)
-        if self.align_bytes <= 0 or self.align_bytes & (self.align_bytes - 1) != 0:
-            raise ValueError("align_bytes must be a positive power of two")
-
-        # Extract shm_name from config.extra_config if available
-        config = kwargs.get("config", None)
-        if config is not None:
-            self.shm_name: Optional[str] = config.get_extra_config_value(
-                "shm_name", None
-            )
-        else:
-            self.shm_name = kwargs.get("shm_name", None)
-
-        self.size = size
-
-        self.buffer = _allocate_cpu_memory(
-            size, self.numa_mapping, self.shm_name, use_hugepages=use_hugepages
-        )
-
-        self._unregistered = False
-
-        self.pin_allocator: MemoryAllocatorInterface
-        if use_paging:
-            assert "shapes" in kwargs, (
-                "shapes must be specified for paged memory allocator"
-            )
-            assert "dtypes" in kwargs, (
-                "dtypes must be specified for paged memory allocator"
-            )
-            assert "fmt" in kwargs, "fmt must be specified for paged memory allocator"
-            self.pin_allocator = PagedTensorMemoryAllocator(
-                tensor=self.buffer,
-                shapes=kwargs["shapes"],
-                dtypes=kwargs["dtypes"],
-                fmt=kwargs["fmt"],
-            )
-        else:
-            self.pin_allocator = TensorMemoryAllocator(
-                self.buffer, align_bytes=self.align_bytes
-            )
-
-        self.host_mem_lock = threading.Lock() if not use_paging else nullcontext()
-
-        self.buffer_allocator = BufferAllocator("cpu")
-
-    @_lmcache_nvtx_annotate
-    def allocate(
-        self,
-        shapes: Union[torch.Size, list[torch.Size]],
-        dtypes: Union[torch.dtype, list[torch.dtype]],
-        fmt: MemoryFormat = MemoryFormat.KV_2LTD,
-        allocator_type: Optional[str] = None,
-    ) -> Optional[MemoryObj]:
-        if fmt == MemoryFormat.BINARY_BUFFER:
-            return self.buffer_allocator.allocate(shapes, dtypes, fmt)
-        elif fmt in [
-            MemoryFormat.KV_2LTD,
-            MemoryFormat.KV_2TD,
-            MemoryFormat.KV_T2D,
-            MemoryFormat.KV_MLA_FMT,
-            MemoryFormat.EC_TD,
-            MemoryFormat.HS_TD,
-        ]:
-            with self.host_mem_lock:
-                obj = self.pin_allocator.allocate(shapes, dtypes, fmt, str(self))
-                if isinstance(obj, TensorMemoryObj):
-                    obj.parent_allocator = self
-                return obj
-        else:
-            raise ValueError(f"Unsupported memory format: {fmt}")
-
-    @_lmcache_nvtx_annotate
-    def batched_allocate(
-        self,
-        shapes: Union[torch.Size, list[torch.Size]],
-        dtypes: Union[torch.dtype, list[torch.dtype]],
-        batch_size: int,
-        fmt: MemoryFormat = MemoryFormat.KV_2LTD,
-        allocator_type: Optional[str] = None,
-    ) -> Optional[List[MemoryObj]]:
-        if fmt == MemoryFormat.BINARY_BUFFER:
-            return self.buffer_allocator.batched_allocate(
-                shapes, dtypes, batch_size, fmt
-            )
-        elif fmt in [
-            MemoryFormat.KV_2LTD,
-            MemoryFormat.KV_2TD,
-            MemoryFormat.KV_T2D,
-            MemoryFormat.KV_MLA_FMT,
-            MemoryFormat.EC_TD,
-            MemoryFormat.HS_TD,
-        ]:
-            with self.host_mem_lock:
-                objs = self.pin_allocator.batched_allocate(
-                    shapes, dtypes, batch_size, fmt, str(self)
-                )
-                if objs is not None:
-                    for obj in objs:
-                        if isinstance(obj, TensorMemoryObj):
-                            obj.parent_allocator = self
-                return objs
-        else:
-            raise ValueError(f"Unsupported memory format: {fmt}")
-
-    @_lmcache_nvtx_annotate
-    def free(self, memory_obj: MemoryObj, allocator_type: Optional[str] = None):
-        fmt = memory_obj.meta.fmt
-        if fmt == MemoryFormat.BINARY_BUFFER:
-            self.buffer_allocator.free(memory_obj)
-        elif fmt in [
-            MemoryFormat.KV_2LTD,
-            MemoryFormat.KV_2TD,
-            MemoryFormat.KV_T2D,
-            MemoryFormat.KV_MLA_FMT,
-            MemoryFormat.EC_TD,
-            MemoryFormat.HS_TD,
-        ]:
-            with self.host_mem_lock:
-                self.pin_allocator.free(memory_obj)
-        else:
-            raise ValueError(f"Unsupported memory format: {fmt}")
-
-    @_lmcache_nvtx_annotate
-    def batched_free(
-        self,
-        memory_objs: List[MemoryObj],
-        allocator_type: Optional[str] = None,
-        update_stats: bool = True,
-    ):
-        if not memory_objs:
-            return
-
-        # NOTE: fmts of all memory_objs should be the same
-        fmt = memory_objs[0].meta.fmt
-        if fmt == MemoryFormat.BINARY_BUFFER:
-            self.buffer_allocator.batched_free(memory_objs)
-        elif fmt in [
-            MemoryFormat.KV_2LTD,
-            MemoryFormat.KV_2TD,
-            MemoryFormat.KV_T2D,
-            MemoryFormat.KV_MLA_FMT,
-            MemoryFormat.EC_TD,
-            MemoryFormat.HS_TD,
-        ]:
-            with self.host_mem_lock:
-                self.pin_allocator.batched_free(memory_objs)
-        else:
-            raise ValueError(f"Unsupported memory format: {fmt}")
-
-    def memcheck(self):
-        with self.host_mem_lock:
-            return self.pin_allocator.memcheck()
-
-    def close(self):
-        if not self._unregistered:
-            if torch_dev.is_available():
-                torch_dev.synchronize()
-            if self.buffer.numel() == 0:
-                return
-            _free_cpu_memory(
-                self.buffer,
-                self.size,
-                self.numa_mapping,
-                self.shm_name,
-                use_hugepages=self.use_hugepages,
-            )
-            self._unregistered = True
-
-    def get_paged_buffers(self) -> Optional[tuple[torch.Tensor, ...]]:
-        """
-        Get the paged buffers for fixed buffer registration.
-
-        Returns:
-            Tuple of paged buffer tensors if using paged allocator, None otherwise.
-            These buffers can be registered with io_uring for true zero copy operations.
-        """
-        if isinstance(self.pin_allocator, PagedTensorMemoryAllocator):
-            return self.pin_allocator.get_paged_buffers()
-        return None
-
-    def __str__(self):
-        return "MixedMemoryAllocator"
-
-
-class DevDaxMemoryAllocator(MemoryAllocatorInterface):
-    """Allocates L1 objects from DRAM and/or an mmap-backed Device-DAX arena.
-
-    The mapped bytes are exposed as a flat CPU ``torch.uint8`` tensor so the
-    existing L1 state machine and tensor slicing logic can run unchanged while
-    the overflow backing storage is the configured Device-DAX device. When a
-    local allocator is provided, local DRAM is tried first and Device-DAX is
-    used as overflow.
-    """
-
-    def __init__(
-        self,
-        size: int,
-        device_path: str,
-        *,
-        local_allocator: MixedMemoryAllocator | None = None,
-        local_size: int = 0,
-        shm_name: str | None = None,
-        align_bytes: int = AddressManager.ALIGN_BYTES,
-    ) -> None:
-        if not device_path:
-            raise ValueError("device_path must be a non-empty string")
-        if size <= 0:
-            raise ValueError("size must be > 0")
-        if local_size < 0:
-            raise ValueError("local_size must be >= 0")
-        if local_size and local_allocator is not None:
-            raise ValueError("local_size cannot be used with local_allocator")
-        if align_bytes <= 0 or align_bytes & (align_bytes - 1) != 0:
-            raise ValueError("align_bytes must be a positive power of two")
-
-        self.size = size
-        self.device_path = device_path
-        self.align_bytes = align_bytes
-        self.local_allocator = local_allocator
-        if local_size:
-            self.local_allocator = MixedMemoryAllocator(
-                local_size,
-                align_bytes=self.align_bytes,
-                shm_name=shm_name,
-            )
-        self.host_mem_lock = threading.Lock()
-        self.buffer_allocator = BufferAllocator("cpu")
-        self.devdax_allocator: TensorMemoryAllocator | None = None
-        self._fd: int | None = None
-        self._mmap_obj: mmap.mmap | None = None
-        self._mmap_buffer: Any | None = None
-        self._unregistered = False
-        self._host_memory_pinned_ptr: int | None = None
-
-        self.devdax_buffer = self._map_devdax()
-        self.devdax_allocator = TensorMemoryAllocator(
-            self.devdax_buffer, align_bytes=self.align_bytes
-        )
-        self.address_manager = self.devdax_allocator.address_manager
-        self._register_cuda_host_memory()
-
-    @property
-    def buffer(self) -> torch.Tensor:
-        if self.local_allocator is not None:
-            return self.local_allocator.buffer
-        return self.devdax_buffer
-
-    def _open_devdax_mapping(
-        self,
-        device_path: str,
-        size: int,
-    ) -> tuple[int, mmap.mmap, Any, torch.Tensor]:
-        fd: int | None = None
-        mmap_obj: mmap.mmap | None = None
-        try:
-            fd = os.open(device_path, os.O_RDWR)
-            capacity = os.fstat(fd).st_size
-            if capacity > 0 and size > capacity:
-                raise RuntimeError(
-                    f"l1 devdax size ({size} bytes) exceeds "
-                    f"{device_path} capacity ({capacity} bytes)"
-                )
-
-            mmap_obj = mmap.mmap(
-                fd,
-                size,
-                flags=mmap.MAP_SHARED,
-                prot=mmap.PROT_READ | mmap.PROT_WRITE,
-            )
-            array_type = ctypes.c_uint8 * size
-            mmap_buffer = array_type.from_buffer(mmap_obj)
-            buffer = torch.frombuffer(mmap_buffer, dtype=torch.uint8)
-            return fd, mmap_obj, mmap_buffer, buffer
-        except Exception:
-            if mmap_obj is not None:
-                mmap_obj.close()
-            if fd is not None:
-                os.close(fd)
-            raise
-
-    def _map_devdax(self) -> torch.Tensor:
-        fd, mmap_obj, mmap_buffer, buffer = self._open_devdax_mapping(
-            self.device_path,
-            self.size,
-        )
-        self._fd = fd
-        self._mmap_obj = mmap_obj
-        self._mmap_buffer = mmap_buffer
-        return buffer
-
-    def _close_devdax_mapping_locked(self) -> None:
-        self._unregister_cuda_host_memory()
-        self.devdax_allocator = None
-        self.devdax_buffer = torch.empty(0, dtype=torch.uint8)
-        self._mmap_buffer = None
-
-        if self._mmap_obj is not None:
-            self._mmap_obj.close()
-            self._mmap_obj = None
-        if self._fd is not None:
-            os.close(self._fd)
-            self._fd = None
-
-    def _register_cuda_host_memory(self) -> None:
-        if not torch_dev.ext.is_pin_supported:
-            return
-        if not torch_dev.ext.pin_memory(self.devdax_buffer.data_ptr(), self.size):
-            logger.warning(
-                "pin_memory failed for Device-DAX L1 mapping; "
-                "falling back to pageable host copies"
-            )
-            return
-        self._host_memory_pinned_ptr = self.devdax_buffer.data_ptr()
-
-    def _unregister_cuda_host_memory(self) -> None:
-        if self._host_memory_pinned_ptr is None:
-            return
-        torch_dev.ext.unpin_memory(self._host_memory_pinned_ptr)
-        self._host_memory_pinned_ptr = None
-
-    def _is_local_obj(self, memory_obj: MemoryObj) -> bool:
-        return (
-            self.local_allocator is not None
-            and memory_obj.parent() is self.local_allocator
-        )
-
-    def _is_devdax_obj(self, memory_obj: MemoryObj) -> bool:
-        return memory_obj.parent() is self
-
-    def _local_available_count(
-        self,
-        shapes: Union[torch.Size, list[torch.Size]],
-        dtypes: Union[torch.dtype, list[torch.dtype]],
-    ) -> int:
-        if self.local_allocator is None:
-            return 0
-        local_pin_allocator = self.local_allocator.pin_allocator
-        assert isinstance(local_pin_allocator, TensorMemoryAllocator)
-        shapes, dtypes = self._adapt_shapes_and_dtypes(shapes, dtypes)
-        unit_raw_size = get_size_bytes(shapes, dtypes)
-        unit_aligned_size = local_pin_allocator.address_manager.compute_aligned_size(
-            unit_raw_size
-        )
-        return local_pin_allocator.address_manager.get_free_size() // unit_aligned_size
-
-    def _local_memory_usage(self) -> tuple[int, int]:
-        if self.local_allocator is None:
-            return 0, 0
-        local_pin_allocator = self.local_allocator.pin_allocator
-        assert isinstance(local_pin_allocator, TensorMemoryAllocator)
-        local_total = local_pin_allocator.address_manager.get_heap_size()
-        local_used = local_total - local_pin_allocator.address_manager.get_free_size()
-        return local_used, local_total
-
-    def _allocate_from_devdax(
-        self,
-        shapes: Union[torch.Size, list[torch.Size]],
-        dtypes: Union[torch.dtype, list[torch.dtype]],
-        fmt: MemoryFormat,
-    ) -> Optional[MemoryObj]:
-        with self.host_mem_lock:
-            assert self.devdax_allocator is not None
-            obj = self.devdax_allocator.allocate(shapes, dtypes, fmt, str(self))
-            if isinstance(obj, TensorMemoryObj):
-                obj.parent_allocator = self
-            return obj
-
-    def _batched_allocate_from_devdax(
-        self,
-        shapes: Union[torch.Size, list[torch.Size]],
-        dtypes: Union[torch.dtype, list[torch.dtype]],
-        batch_size: int,
-        fmt: MemoryFormat,
-    ) -> Optional[List[MemoryObj]]:
-        with self.host_mem_lock:
-            assert self.devdax_allocator is not None
-            objs = self.devdax_allocator.batched_allocate(
-                shapes, dtypes, batch_size, fmt, str(self)
-            )
-            if objs is not None:
-                for obj in objs:
-                    if isinstance(obj, TensorMemoryObj):
-                        obj.parent_allocator = self
-            return objs
-
-    def _free_devdax_obj(self, memory_obj: MemoryObj) -> None:
-        with self.host_mem_lock:
-            assert self.devdax_allocator is not None
-            self.devdax_allocator.free(memory_obj)
-            if isinstance(memory_obj, TensorMemoryObj):
-                memory_obj.raw_data = torch.empty(0, dtype=torch.uint8)
-
-    def _batched_free_devdax_objs(
-        self,
-        memory_objs: List[MemoryObj],
-        update_stats: bool,
-    ) -> None:
-        with self.host_mem_lock:
-            assert self.devdax_allocator is not None
-            self.devdax_allocator.batched_free(memory_objs, update_stats=update_stats)
-            for memory_obj in memory_objs:
-                if isinstance(memory_obj, TensorMemoryObj):
-                    memory_obj.raw_data = torch.empty(0, dtype=torch.uint8)
-
-    @_lmcache_nvtx_annotate
-    def allocate(
-        self,
-        shapes: Union[torch.Size, list[torch.Size]],
-        dtypes: Union[torch.dtype, list[torch.dtype]],
-        fmt: MemoryFormat = MemoryFormat.KV_2LTD,
-        allocator_type: Optional[str] = None,
-    ) -> Optional[MemoryObj]:
-        if fmt == MemoryFormat.BINARY_BUFFER:
-            return self.buffer_allocator.allocate(shapes, dtypes, fmt)
-        elif fmt in [
-            MemoryFormat.KV_2LTD,
-            MemoryFormat.KV_2TD,
-            MemoryFormat.KV_T2D,
-            MemoryFormat.KV_MLA_FMT,
-            MemoryFormat.EC_TD,
-        ]:
-            if self.local_allocator is not None:
-                obj = self.local_allocator.allocate(shapes, dtypes, fmt, str(self))
-                if obj is not None:
-                    return obj
-            if self.devdax_allocator is None:
-                return None
-            return self._allocate_from_devdax(shapes, dtypes, fmt)
-        else:
-            raise ValueError(f"Unsupported memory format: {fmt}")
-
-    @_lmcache_nvtx_annotate
-    def batched_allocate(
-        self,
-        shapes: Union[torch.Size, list[torch.Size]],
-        dtypes: Union[torch.dtype, list[torch.dtype]],
-        batch_size: int,
-        fmt: MemoryFormat = MemoryFormat.KV_2LTD,
-        allocator_type: Optional[str] = None,
-    ) -> Optional[List[MemoryObj]]:
-        if fmt == MemoryFormat.BINARY_BUFFER:
-            return self.buffer_allocator.batched_allocate(
-                shapes, dtypes, batch_size, fmt
-            )
-        elif fmt in [
-            MemoryFormat.KV_2LTD,
-            MemoryFormat.KV_2TD,
-            MemoryFormat.KV_T2D,
-            MemoryFormat.KV_MLA_FMT,
-            MemoryFormat.EC_TD,
-        ]:
-            local_objs: list[MemoryObj] = []
-            if self.local_allocator is not None:
-                local_count = min(
-                    batch_size, self._local_available_count(shapes, dtypes)
-                )
-                if local_count:
-                    local_objs = (
-                        self.local_allocator.batched_allocate(
-                            shapes, dtypes, local_count, fmt, str(self)
-                        )
-                        or []
-                    )
-
-            remaining = batch_size - len(local_objs)
-            if remaining == 0:
-                return local_objs
-
-            if self.devdax_allocator is None:
-                if local_objs and self.local_allocator is not None:
-                    self.local_allocator.batched_free(local_objs, update_stats=False)
-                return None
-
-            dax_objs = self._batched_allocate_from_devdax(
-                shapes, dtypes, remaining, fmt
-            )
-            if dax_objs is None:
-                if local_objs and self.local_allocator is not None:
-                    self.local_allocator.batched_free(local_objs, update_stats=False)
-                return None
-            return local_objs + dax_objs
-        else:
-            raise ValueError(f"Unsupported memory format: {fmt}")
-
-    @_lmcache_nvtx_annotate
-    def free(self, memory_obj: MemoryObj, allocator_type: Optional[str] = None) -> None:
-        fmt = memory_obj.meta.fmt
-        if fmt == MemoryFormat.BINARY_BUFFER:
-            self.buffer_allocator.free(memory_obj)
-        elif fmt in [
-            MemoryFormat.KV_2LTD,
-            MemoryFormat.KV_2TD,
-            MemoryFormat.KV_T2D,
-            MemoryFormat.KV_MLA_FMT,
-            MemoryFormat.EC_TD,
-        ]:
-            if self._is_local_obj(memory_obj):
-                assert self.local_allocator is not None
-                self.local_allocator.free(memory_obj)
-                return
-            if self._is_devdax_obj(memory_obj):
-                self._free_devdax_obj(memory_obj)
-                return
-            raise ValueError("Memory object does not belong to DevDaxMemoryAllocator")
-        else:
-            raise ValueError(f"Unsupported memory format: {fmt}")
-
-    @_lmcache_nvtx_annotate
-    def batched_free(
-        self,
-        memory_objs: List[MemoryObj],
-        allocator_type: Optional[str] = None,
-        update_stats: bool = True,
-    ) -> None:
-        if not memory_objs:
-            return
-
-        fmt = memory_objs[0].meta.fmt
-        if fmt == MemoryFormat.BINARY_BUFFER:
-            self.buffer_allocator.batched_free(memory_objs)
-        elif fmt in [
-            MemoryFormat.KV_2LTD,
-            MemoryFormat.KV_2TD,
-            MemoryFormat.KV_T2D,
-            MemoryFormat.KV_MLA_FMT,
-            MemoryFormat.EC_TD,
-        ]:
-            local_objs = [
-                memory_obj
-                for memory_obj in memory_objs
-                if self._is_local_obj(memory_obj)
-            ]
-            devdax_objs = [
-                memory_obj
-                for memory_obj in memory_objs
-                if self._is_devdax_obj(memory_obj)
-            ]
-            if len(local_objs) + len(devdax_objs) != len(memory_objs):
-                raise ValueError(
-                    "One or more memory objects do not belong to DevDaxMemoryAllocator"
-                )
-            if local_objs:
-                assert self.local_allocator is not None
-                self.local_allocator.batched_free(local_objs, update_stats=update_stats)
-            if devdax_objs:
-                self._batched_free_devdax_objs(devdax_objs, update_stats=update_stats)
-        else:
-            raise ValueError(f"Unsupported memory format: {fmt}")
-
-    def memcheck(self) -> bool:
-        local_ok = True
-        if self.local_allocator is not None:
-            local_ok = self.local_allocator.memcheck()
-        with self.host_mem_lock:
-            if self.devdax_allocator is None:
-                return local_ok
-            return local_ok and self.devdax_allocator.memcheck()
-
-    def get_memory_usage(self) -> tuple[int, int]:
-        local_used, local_total = self._local_memory_usage()
-
-        if self.devdax_allocator is None:
-            return local_used, local_total
-        dax_total = self.devdax_allocator.address_manager.get_heap_size()
-        dax_used = dax_total - self.devdax_allocator.address_manager.get_free_size()
-        return local_used + dax_used, local_total + dax_total
-
-    def close(self) -> None:
-        if self._unregistered:
-            return
-        if torch_dev.is_available():
-            torch_dev.synchronize()
-
-        with self.host_mem_lock:
-            if (
-                self.devdax_allocator is not None
-                and self.devdax_allocator.num_active_allocations > 0
-            ):
-                raise BufferError(
-                    "cannot close DevDaxMemoryAllocator with active allocations"
-                )
-            self._close_devdax_mapping_locked()
-
-        if self.local_allocator is not None:
-            self.local_allocator.close()
-            self.local_allocator = None
-        self._unregistered = True
-
-    def __del__(self) -> None:
-        try:
-            self.close()
-        except Exception:
-            pass
-
-    def __str__(self) -> str:
-        return "DevDaxMemoryAllocator"
-
-
-class GPUMemoryAllocator(MemoryAllocatorInterface):
-    """Allocates memory in the pre-allocated GPU memory."""
-
-    def __init__(
-        self,
-        size: int,
-        device=torch_device_type,
-        align_bytes: Optional[int] = None,
-        use_paging: bool = False,
-        **kwargs,
-    ):
-        """
-        :param int size: The size of the GPU memory in bytes.
-        :param Optional[int] align_bytes: The byte alignment for allocations.
-        """
-        if not torch_dev.is_available():
-            device = "cpu"
-
-        self.tensor = torch.empty(size, dtype=torch.uint8, device=device)
-
-        self.allocator: MemoryAllocatorInterface
-        if use_paging:
-            assert "shapes" in kwargs, (
-                "shapes must be specified for paged memory allocator"
-            )
-            assert "dtypes" in kwargs, (
-                "dtypes must be specified for paged memory allocator"
-            )
-            assert "fmt" in kwargs, "fmt must be specified for paged memory allocator"
-            self.allocator = PagedTensorMemoryAllocator(
-                tensor=self.tensor,
-                shapes=kwargs["shapes"],
-                dtypes=kwargs["dtypes"],
-                fmt=kwargs["fmt"],
-            )
-        else:
-            kwargs = {}
-            if align_bytes is not None:
-                kwargs["align_bytes"] = align_bytes
-            self.allocator = TensorMemoryAllocator(self.tensor, **kwargs)
-
-        self.device_mem_lock = threading.Lock() if not use_paging else nullcontext()
-
-    @_lmcache_nvtx_annotate
-    def allocate(
-        self,
-        shapes: Union[torch.Size, list[torch.Size]],
-        dtypes: Union[torch.dtype, list[torch.dtype]],
-        fmt: MemoryFormat = MemoryFormat.KV_2LTD,
-        allocator_type: Optional[str] = None,
-    ) -> Optional[MemoryObj]:
-        with self.device_mem_lock:
-            return self.allocator.allocate(shapes, dtypes, fmt, str(self))
-
-    @_lmcache_nvtx_annotate
-    def batched_allocate(
-        self,
-        shapes: Union[torch.Size, list[torch.Size]],
-        dtypes: Union[torch.dtype, list[torch.dtype]],
-        batch_size: int,
-        fmt: MemoryFormat = MemoryFormat.KV_2LTD,
-        allocator_type: Optional[str] = None,
-    ) -> Optional[List[MemoryObj]]:
-        with self.device_mem_lock:
-            return self.allocator.batched_allocate(
-                shapes, dtypes, batch_size, fmt, str(self)
-            )
-
-    def free(self, memory_obj: MemoryObj, allocator_type: Optional[str] = None):
-        with self.device_mem_lock:
-            self.allocator.free(memory_obj)
-
-    def batched_free(
-        self,
-        memory_objs: List[MemoryObj],
-        allocator_type: Optional[str] = None,
-        update_stats: bool = True,
-    ):
-        with self.device_mem_lock:
-            self.allocator.batched_free(memory_objs)
-
-    def memcheck(self):
-        with self.device_mem_lock:
-            return self.allocator.memcheck()
-
-    def __str__(self):
-        return "GPUMemoryAllocator"
-
-
-class AdHocMemoryAllocator(MemoryAllocatorInterface):
-    """
-    AdHocMemoryAllocator is a simple allocator that does not actually
-    allocate memory. It is used for testing purposes only.
-    """
-
-    def __init__(self, device: str = "cpu"):
-        """
-        :param str device: The device of the ad hoc memory allocator.
-        """
-        if not torch_dev.is_available():
-            self.device = "cpu"
-        else:
-            self.device = device
-
-    @_lmcache_nvtx_annotate
-    def allocate(
-        self,
-        shapes: Union[torch.Size, list[torch.Size]],
-        dtypes: Union[torch.dtype, list[torch.dtype]],
-        fmt: MemoryFormat = MemoryFormat.KV_2LTD,
-        allocator_type: Optional[str] = None,
-    ) -> Optional[MemoryObj]:
-        """
-        Returns a dummy MemoryObj for testing purposes.
-        """
-        shapes, dtypes = self._adapt_shapes_and_dtypes(shapes, dtypes)
-        size = get_size_bytes(shapes, dtypes)
-
-        # Return a dummy object with no actual memory allocation
-        return TensorMemoryObj(
-            raw_data=torch.empty(
-                torch.Size([size]), dtype=torch.uint8, device=self.device
-            ),
-            metadata=MemoryObjMetadata(
-                shape=shapes[0],
-                dtype=dtypes[0],
-                address=0,
-                phy_size=0,
-                ref_count=1,
-                pin_count=0,
-                fmt=fmt,
-                shapes=shapes,
-                dtypes=dtypes,
-            ),
-            parent_allocator=self,
-        )
-
-    @_lmcache_nvtx_annotate
-    def batched_allocate(
-        self,
-        shapes: Union[torch.Size, list[torch.Size]],
-        dtypes: Union[torch.dtype, list[torch.dtype]],
-        batch_size: int,
-        fmt: MemoryFormat = MemoryFormat.KV_2LTD,
-        allocator_type: Optional[str] = None,
-    ) -> Optional[List[MemoryObj]]:
-        raise NotImplementedError(
-            "Batched allocation is not supported in AdHocMemoryAllocator"
-        )
-
-    def free(self, memory_obj: MemoryObj, allocator_type: Optional[str] = None):
-        pass
-
-    def batched_free(
-        self,
-        memory_objs: List[MemoryObj],
-        allocator_type: Optional[str] = None,
-        update_stats: bool = True,
-    ):
-        pass
-
-    def ref_count_up(self, memory_obj: MemoryObj):
-        pass
-
-    def ref_count_down(self, memory_obj: MemoryObj):
-        pass
-
-    def get_ref_count(self, memory_obj: MemoryObj):
-        return 0
-
-    def memcheck(self):
-        return True
-
-    def __str__(self):
-        return "AdHocMemoryAllocator"
-
-
-class CuFileMemoryAllocator(GPUMemoryAllocator):
-    def __init__(self, size: int, device=None):
-        # HACK(Jiayi): cufile import is buggy on some hardware
-        # (e.g., without GPUDirect), so it's temporarily put here.
-        # Third Party
-        from cufile.bindings import cuFileBufDeregister, cuFileBufRegister
-
-        self.cuFileBufDeregister = cuFileBufDeregister
-        if device is None:
-            # TODO(Serapheim): Ideally we'd get the device from the upper
-            # layer - for now just use the current device.
-            if torch_dev.is_available():
-                device = f"{torch_device_type}:{torch_dev.current_device()}"
-            else:
-                device = "cpu:0"
-        super().__init__(size, device, align_bytes=4096)
-        self.base_pointer = self.tensor.data_ptr()
-        cuFileBufRegister(ctypes.c_void_p(self.base_pointer), size, flags=0)
-
-    def __del__(self):
-        self.cuFileBufDeregister(ctypes.c_void_p(self.base_pointer))
-
-    def __str__(self):
-        return "CuFileMemoryAllocator"
-
-
-class HipFileMemoryAllocator(GPUMemoryAllocator):
-    def __init__(self, size: int, device=None):
-        # HACK: hipfile import is placed here to avoid import errors on
-        # hardware without GPUDirect Storage / hipFile support.
-        # Third Party
-        from hipfile import Buffer
-
-        if device is None:
-            if torch_dev.is_available():
-                # TODO: On ROCm, PyTorch still uses the CUDA API internally
-                device = f"{torch_device_type}:{torch_dev.current_device()}"
-            else:
-                device = "cpu:0"
-
-        super().__init__(size, device, align_bytes=4096)
-        self.base_pointer = self.tensor.data_ptr()
-        self.hipfile_buffer = Buffer(self.base_pointer, size, flags=0)
-        self.hipfile_buffer.register()
-
-    def __del__(self):
-        try:
-            self.hipfile_buffer.deregister()
-        except Exception:
-            pass
-
-    def __str__(self):
-        return "HipFileMemoryAllocator"
-
-
-class PagedCpuGpuMemoryAllocator(MemoryAllocatorInterface):
-    """
-    Paged Memory Allocator for both CPU and GPU memory.
-    This is a paged memory allocator for PD and P2P sharing
-    when NIXL is enabled as NIXL relies on the paging abstraction.
-    """
-
-    def __init__(self):
-        pass
-
-    def init_gpu_memory_allocator(
-        self,
-        size: int,
-        shapes: list[torch.Size],
-        dtypes: list[torch.dtype],
-        fmt: MemoryFormat = MemoryFormat.KV_2LTD,
-        device: str = torch_device_type,
-    ):
-        self.gpu_buffer = torch.empty(
-            size,
-            dtype=torch.uint8,
-            device=device,
-        )
-        self.gpu_allocator = PagedTensorMemoryAllocator(
-            self.gpu_buffer,
-            shapes,
-            dtypes,
-            fmt,
-        )
-
-    def init_cpu_memory_allocator(
-        self,
-        size: int,
-        shapes: list[torch.Size],
-        dtypes: list[torch.dtype],
-        fmt: MemoryFormat = MemoryFormat.KV_2LTD,
-        numa_mapping: Optional[NUMAMapping] = None,
-    ):
-        self.cpu_buffer = _allocate_cpu_memory(size, numa_mapping)
-        self.cpu_allocator = PagedTensorMemoryAllocator(
-            self.cpu_buffer,
-            shapes,
-            dtypes,
-            fmt,
-        )
-        self.align_bytes = self.cpu_allocator.align_bytes
-
-    def allocate(
-        self,
-        shapes: Union[torch.Size, list[torch.Size]],
-        dtypes: Union[torch.dtype, list[torch.dtype]],
-        fmt: MemoryFormat = MemoryFormat.UNDEFINED,
-        allocator_type: Optional[str] = "cpu",
-    ) -> Optional[MemoryObj]:
-        if allocator_type == "gpu":
-            return self.gpu_allocator.allocate(shapes, dtypes, fmt)
-        elif allocator_type == "cpu":
-            return self.cpu_allocator.allocate(shapes, dtypes, fmt)
-        else:
-            raise ValueError(f"Unsupported allocator type: {allocator_type}")
-
-    def batched_allocate(
-        self,
-        shapes: Union[torch.Size, list[torch.Size]],
-        dtypes: Union[torch.dtype, list[torch.dtype]],
-        batch_size: int,
-        fmt: MemoryFormat = MemoryFormat.UNDEFINED,
-        allocator_type: Optional[str] = "gpu",
-    ) -> Optional[List[MemoryObj]]:
-        if allocator_type == "gpu":
-            return self.gpu_allocator.batched_allocate(shapes, dtypes, batch_size, fmt)
-        elif allocator_type == "cpu":
-            return self.cpu_allocator.batched_allocate(shapes, dtypes, batch_size, fmt)
-        else:
-            raise ValueError(f"Unsupported allocator type: {allocator_type}")
-
-    def free(self, memory_obj: MemoryObj, allocator_type: Optional[str] = "cpu"):
-        if allocator_type == "gpu":
-            self.gpu_allocator.free(memory_obj)
-        elif allocator_type == "cpu":
-            self.cpu_allocator.free(memory_obj)
-        else:
-            raise ValueError(f"Unsupported allocator type: {allocator_type}")
-
-    def batched_free(
-        self,
-        memory_objs: List[MemoryObj],
-        allocator_type: Optional[str] = None,
-        update_stats: bool = True,
-    ):
-        if allocator_type == "gpu":
-            self.gpu_allocator.batched_free(memory_objs, update_stats=update_stats)
-        elif allocator_type == "cpu":
-            self.cpu_allocator.batched_free(memory_objs, update_stats=update_stats)
-        else:
-            raise ValueError(f"Unsupported allocator type: {allocator_type}")
-
-    def __str__(self):
-        return "PDMemoryAllocator"
+# First Party
+from lmcache.v1.memory_allocators import (  # noqa: E402
+    AdHocMemoryAllocator,
+    BufferAllocator,
+    CuFileMemoryAllocator,
+    DevDaxMemoryAllocator,
+    GPUMemoryAllocator,
+    HipFileMemoryAllocator,
+    HostMemoryAllocator,
+    MixedMemoryAllocator,
+    PagedCpuGpuMemoryAllocator,
+    PagedTensorMemoryAllocator,
+    PinMemoryAllocator,
+    TensorMemoryAllocator,
+)
+
+__all__ = [
+    "AdHocMemoryAllocator",
+    "AddressManager",
+    "BufferAllocator",
+    "BytesBufferMemoryObj",
+    "CuFileMemoryAllocator",
+    "DevDaxMemoryAllocator",
+    "FreeBlock",
+    "GDSMemoryObject",
+    "GPUMemoryAllocator",
+    "HipFileMemoryAllocator",
+    "HostMemoryAllocator",
+    "MemoryAllocatorInterface",
+    "MemoryFormat",
+    "MemoryObj",
+    "MemoryObjMetadata",
+    "MixedMemoryAllocator",
+    "PagedCpuGpuMemoryAllocator",
+    "PagedTensorMemoryAllocator",
+    "PinMemoryAllocator",
+    "TensorMemoryAllocator",
+    "TensorMemoryObj",
+    "torch_device_type",
+]
